Optimize day 6 part 2.
An order of magnitude speed up by marking corners, rather than all visited
cells. To calculate all the visited cells we just interpolate between the
corners and remove duplicates. We know we're in a loop if we hit the same
corner in the same direction as we've been there before.